Fix two errors found by PVS-Studio

+++ tests/msgpack.spec.tests/ReadOnlySequence/MultipleSegments/TryFillSpan.cs
@@ -38,7 +38,7 @@ namespace ProGaudi.MsgPack.Tests.ReadOnlySequence.MultipleSegments
         [Fact]
         public void TryFillSpanNonByteEmpty()
         {
-            var empty = System.Array.Empty<byte>();
+            var empty = System.Array.Empty<int>();
             empty.ToMultipleSegments().TryFillSpan(empty).ShouldBeTrue();
         }
     }
